+++ include/hegel/hegel.h
@@ -20,7 +20,7 @@
  * FetchContent_Declare(
  *     hegel
  *     GIT_REPOSITORY https://github.com/hegeldev/hegel-cpp.git
- *     GIT_TAG v0.3.0
+ *     GIT_TAG v0.3.3
  * )
  * FetchContent_MakeAvailable(hegel)
  *